backend/requirements.txt
Flask
Flask-Cors
Pillow
piexif
reportlab
werkzeug
